+++ example/src/App.svelte
@@ -2,4 +2,13 @@
 	import Typewriter from 'Typewriter.svelte'
 </script>
 
-<Typewriter loopRandom>Hello World!</Typewriter>
+<Typewriter scramble on:done={() => console.log('Hello World!')}>
+	<div>some text...</div>
+	<div>
+		Foo
+		<span>
+      <a href="user">some link</a> <a href="user">yet another link</a>
+    </span>
+		Bar
+	</div>
+</Typewriter>