+++ CHANGELOG.md
@@ -2,11 +2,12 @@
 
 All notable changes to this project will be documented in this file. Dates are displayed in UTC.
 
-#### [Unreleased](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.1...HEAD)
+#### [v2.4.2](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.1...v2.4.2)
 
+- fix(cascade): properly animate nested elements [`#27`](https://github.com/henriquehbr/svelte-typewriter/pull/27)
 - chore: bump dependencies [`798ff61`](https://github.com/henriquehbr/svelte-typewriter/commit/798ff61ca834125a69b2dbd36fac853ae9e09d7a)
 - fix(cascade): properly animate nested elements (#25) [`519ce6c`](https://github.com/henriquehbr/svelte-typewriter/commit/519ce6c9594125adb81b72d15352c7ad6639cb20)
-- fix(scramble): adjust to work with "getElements" changes (#25) [`1eba996`](https://github.com/henriquehbr/svelte-typewriter/commit/1eba99659b74e9360e8d2d89046b427493fc05ab)
+- fix(scramble): adjust to work with "getElements" changes (#25) [`ae7b457`](https://github.com/henriquehbr/svelte-typewriter/commit/ae7b4570d41ef8069fa90f3cd87e1a9ab99f78d8)
 
 #### [v2.4.1](https://github.com/henriquehbr/svelte-typewriter/compare/v2.4.0...v2.4.1)